requirements.txt
# Backend dependencies for Portfolio WebApp
# Deploy to Railway: Set DATABASE_URL, OWNER_PASSWORD, SECRET_KEY, FRONTEND_ORIGIN, MEDIA_STORAGE env vars

fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Pillow==10.2.0
aiofiles==23.2.1
python-magic==0.4.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
